+++ app/src/components/ParticipantWindow/fragments/ScreenPresenterVideo.tsx
@@ -5,19 +5,15 @@ import { styled } from '@mui/material';
 import { MediaSessionType, ParticipantId, VideoSetting } from '@opentalk/common';
 import React, { useState, useMemo } from 'react';
 
-import NameTile from '../../../commonComponents/NameTile/NameTile';
 import { useAppSelector } from '../../../hooks';
 import { selectQualityCap } from '../../../store/slices/mediaSlice';
-import { selectSubscriberById } from '../../../store/slices/mediaSubscriberSlice';
-import { selectParticipantAvatarUrl, selectParticipantName } from '../../../store/slices/participantsSlice';
+import { selectSubscriberStateById } from '../../../store/slices/mediaSubscriberSlice';
 import { AvatarContainer } from './AvatarContainer';
 import { PresenterVideoPosition } from './ParticipantVideo';
 import { PresenterOverlay } from './PresenterOverlay';
 import RemoteVideo from './RemoteVideo';
 
-const SharedPresenterVideo = styled('div', {
-  shouldForwardProp: (prop) => prop !== 'fullscreenMode',
-})<{ fullscreenMode: boolean | undefined }>(({ theme, fullscreenMode }) => ({
+const SharedPresenterVideo = styled('div')(({ theme }) => ({
   display: 'flex',
   justifyContent: 'center',
   width: '18%',
@@ -33,7 +29,7 @@ const SharedPresenterVideo = styled('div', {
   },
   '&.upperRight': {
     top: theme.spacing(2),
-    right: fullscreenMode ? theme.spacing(8) : theme.spacing(2),
+    right: theme.spacing(2),
   },
   '&.bottomRight': {
     bottom: theme.spacing(2),
@@ -50,28 +46,23 @@ interface ScreenPresenterVideoProps {
   participantId: ParticipantId;
   isVideoPinned: boolean;
   videoPosition: PresenterVideoPosition;
-  isFullscreenMode?: boolean;
   togglePin: () => void;
   changeVideoPosition: () => void;
   isThumbnail?: boolean;
 }
 
 const ScreenPresenterVideo = React.forwardRef<HTMLDivElement, ScreenPresenterVideoProps>(
-  (
-    { participantId, isFullscreenMode, isVideoPinned, togglePin, videoPosition, changeVideoPosition, isThumbnail },
-    ref
-  ) => {
+  ({ participantId, isVideoPinned, togglePin, videoPosition, changeVideoPosition, isThumbnail }, ref) => {
     const videoDescriptor = useMemo(() => ({ participantId, mediaType: MediaSessionType.Video }), [participantId]);
-    const videoSubscriber = useAppSelector(selectSubscriberById(videoDescriptor));
-    const displayName = useAppSelector(selectParticipantName(participantId));
+    const { active, limit } = useAppSelector(selectSubscriberStateById(videoDescriptor, 'video'));
     const [mouseOver, setMouseOver] = useState<boolean>(false);
     const qualityCap = useAppSelector(selectQualityCap);
+    const showVideo = active && limit !== VideoSetting.Off && qualityCap !== VideoSetting.Off;
 
     return (
       <SharedPresenterVideo
         ref={ref}
         className={videoPosition}
-        fullscreenMode={isFullscreenMode}
         data-testid="sharedPresenterVideo"
         onMouseEnter={() => setMouseOver(true)}
         onMouseLeave={() => setMouseOver(false)}
@@ -80,20 +71,12 @@ const ScreenPresenterVideo = React.forwardRef<HTMLDivElement, ScreenPresenterVid
           <PresenterOverlay
             videoPosition={videoPosition}
             participantId={participantId}
-            fullscreenMode={isFullscreenMode}
             isVideoPinned={isVideoPinned}
             togglePin={togglePin}
             changeVideoPosition={changeVideoPosition}
           />
         )}
-        {videoSubscriber?.video && videoSubscriber?.limit !== VideoSetting.Off && qualityCap !== VideoSetting.Off ? (
-          <RemoteVideo descriptor={videoDescriptor} />
-        ) : (
-          <AvatarContainer participantId={participantId} />
-        )}
-        {isFullscreenMode && (
-          <NameTile displayName={displayName || ''} audioOn={!!videoSubscriber?.audio} className="positionBottom" />
-        )}
+        {showVideo ? <RemoteVideo descriptor={videoDescriptor} /> : <AvatarContainer participantId={participantId} />}
       </SharedPresenterVideo>
     );
   }